// framework/src/interfaces/IdeveloperMode.ts
/*
 * File generated by Interface generator (dotup.dotup-vscode-interface-generator)
 * Date: 2022-01-19 12:10:20 
*/
import { IFramework } from "./IFramework";
export interface IDeveloperMode {
    as: IFramework;
    devBarHTMLTemplate: string;
    debugMode: boolean;
    devMode: boolean;
    dragging: boolean;
    devModeWindowState: "open" | "closed";
    open():void
    initialize(): any;
    validateModes(): void;
    implementDevMode(): any;
    createDevModeSlider(): void;
}